:sparkles: 닉네임, 기본 동네 설정 화면 ViewModel
